+++ src/Components/App/App.css
@@ -32,5 +32,11 @@ p
 
 }
 
+.borderBox
+{
+  border-style: solid;
+  padding: 12px;
+  box-shadow:inset 0px 0px 0px 10px #fff;
 
+}
