+++ src/Components/App/App.js
@@ -2,11 +2,11 @@ import React, { useEffect } from 'react';
 import ReactGA from 'react-ga'
 
 import NavBar from '../NavBar/NavBar';
-import Home from '../Pages/Home/Home';
-import Skills from '../Pages/Skills/Skills';
-import ProjectPage from '../Pages/Projects/ProjectPage';
-import Education from '../Pages/Education/Education';
-import Work from '../Pages/Work/Work';
+import Home from '../Sections/Home/Home';
+import AboutMe from '../Sections/AboutMe/AboutMe';
+import ProjectPage from '../Sections/Projects/ProjectPage';
+import Work from '../Sections/Work/Work';
+import Contact from '../Sections/Contact/Contact'
 import './App.css';
 
 // Top Component of the Web app that displays the nav bar, and the active page/component
@@ -14,52 +14,6 @@ import './App.css';
 // (App)
 class App extends React.Component 
 {
-  constructor(props)
-  {
-    super(props);
-
-    this.updateActiveComponent = this.updateActiveComponent.bind(this);
-
-    // Initially have it set to the Home page
-    this.state = {activeComponent: <Home/> };
-  }
-
-  // Function that updates and switch the active page
-  updateActiveComponent(id)
-  {
-    let component;
-
-    // Switch based on the className that is passed back as id
-    switch(id)
-    {
-      case "Home":
-        component = <Home/>
-        break;
-
-      case "Education":
-        component = <Education/>;
-        break;
-    
-      case "Skills":
-        component = <Skills/>;
-        break;
-
-      case "Projects":
-        component = <ProjectPage/>;
-        break;
-
-      case "Work Experience":
-        component = <Work/>;
-        break;
-
-      default:
-        component = <Home/>
-    }
-
-    // Set the new state with teh active component and it's name
-    this.setState({activeComponent: component});
-  }
-
   render()
   {
     return (
@@ -67,8 +21,12 @@ class App extends React.Component
         {/* The font for the website */}
         <link href="https://fonts.googleapis.com/css2?family=Berkshire+Swash&display=swap" rel="stylesheet"/>
 
-        <NavBar class="navBar" onClick={this.updateActiveComponent} />
-        {this.state.activeComponent}
+        <NavBar class="navBar"/> 
+        <Home />
+        <AboutMe />
+        <ProjectPage />
+        <Work />
+        <Contact />
       </div>
     );
   }
