+++ src/Components/Sections/Projects/ProjectBox.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactGA from 'react-ga'
-import '../Projects/Project.css'
+import './Project.css'
 import ProjectTitleCard from './ProjectTitleCard';
 import ProjectDetailsCard from './ProjectDetailsCard'
